+++ build.ps1
@@ -72,6 +72,7 @@ Remove-Item -Recurse $artifacts -ErrorAction Ignore
 exec dotnet msbuild /t:UpdateCiSettings @MSBuildArgs
 exec dotnet build --configuration $Configuration '-warnaserror:CS1591' @MSBuildArgs
 exec dotnet pack --no-restore --no-build --configuration $Configuration -o $artifacts @MSBuildArgs
+exec dotnet build --configuration $Configuration "$PSScriptRoot/docs/samples/samples.sln"
 
 [string[]] $testArgs=@()
 if ($PSVersionTable.PSEdition -eq 'Core' -and -not $IsWindows) {