+++ src/CommandLineUtils/IO/PhysicalConsole.cs
@@ -4,18 +4,28 @@
 
 using System;
 using System.IO;
+using System.Threading;
+using McMaster.Extensions.CommandLineUtils.Internal;
 
 namespace McMaster.Extensions.CommandLineUtils
 {
     /// <summary>
     /// An implementation of <see cref="IConsole"/> that wraps <see cref="System.Console"/>.
     /// </summary>
-    public class PhysicalConsole : IConsole
+    public class PhysicalConsole : IConsole, ICancellationTokenProvider
     {
+        private readonly CancellationTokenSource _cancelKeyPressed;
+
         /// <summary>
         /// A shared instance of <see cref="PhysicalConsole"/>.
         /// </summary>
         public static IConsole Singleton { get; } = new PhysicalConsole();
+
+        private PhysicalConsole()
+        {
+            _cancelKeyPressed = new CancellationTokenSource();
+            Console.CancelKeyPress += (_, __) => _cancelKeyPressed.Cancel();
+        }
 
         /// <summary>
         /// <see cref="Console.CancelKeyPress"/>.
@@ -73,6 +83,8 @@ namespace McMaster.Extensions.CommandLineUtils
             get => Console.BackgroundColor;
             set => Console.BackgroundColor = value;
         }
+
+        CancellationToken ICancellationTokenProvider.Token => _cancelKeyPressed.Token;
 
         /// <summary>
         /// <see cref="Console.ResetColor"/>.