+++ src/CommandLineUtils/Internal/ICancellationTokenProvider.cs
@@ -1,0 +1,12 @@
+// Copyright (c) Nate McMaster.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System.Threading;
+
+namespace McMaster.Extensions.CommandLineUtils.Internal
+{
+    internal interface ICancellationTokenProvider
+    {
+        CancellationToken Token { get; }
+    }
+}